[not compiling] moved code around for binary tree
Also created builder for the tree.
