+++ src/main.rs
@@ -1,25 +1,3 @@
-//////////////////////////////////////////////////////////////////////////////////////////////////////
-// use std::str::FromStr;                                                                           //
-//                                                                                                  //
-// use dapol::{NdmSmt, User, D256, UserId};                                                         //
-//                                                                                                  //
-// fn main() {                                                                                      //
-//     let user1 = User{                                                                            //
-//         liability: 10u64,                                                                        //
-//         id: UserId::from_str("user1 ID").unwrap(),                                               //
-//     };                                                                                           //
-//                                                                                                  //
-//     let master_secret: D256 = D256::from(3u64);                                                  //
-//     let salt_b: D256 = D256::from(5u64);                                                         //
-//     let salt_s: D256 = D256::from(7u64);                                                         //
-//     let height: u8 = 5u8;                                                                        //
-//     let users: Vec<User> = vec![user1];                                                          //
-//     let ndsmt = NdmSmt::new(master_secret, salt_b, salt_s, height, users).unwrap();              //
-//                                                                                                  //
-//     let proof = ndsmt.generate_inclusion_proof(&UserId::from_str("user1 ID").unwrap()).unwrap(); //
-//     println!("{:?}", proof);                                                                     //
-// }                                                                                                //
-//////////////////////////////////////////////////////////////////////////////////////////////////////
 use std::str::FromStr;
 
 use dapol::{NdmSmt, User, UserId, D256};
@@ -54,7 +32,7 @@ fn new() {
     let tree_height = 32;
     // let num_leaves: usize = 2usize.pow(27); // 134M
     let num_leaves: usize = 2usize.pow(23); // 8.4M
-    // let num_leaves: usize = 2usize.pow(17);
+                                            // let num_leaves: usize = 2usize.pow(17);
 
     let users = build_item_list_new(num_leaves, tree_height);
 
@@ -64,8 +42,8 @@ fn new() {
 
     let ndsmt = NdmSmt::new(master_secret, salt_b, salt_s, tree_height as u8, users).unwrap();
 
-    // let proof = ndsmt.generate_inclusion_proof(&UserId::from_str("user1 ID").unwrap()).unwrap();
-    // println!("{:?}", proof);
+    // let proof = ndsmt.generate_inclusion_proof(&UserId::from_str("user1
+    // ID").unwrap()).unwrap(); println!("{:?}", proof);
 }
 
 fn old() {
